migrate plugin test to phpunit
